Move PyperfResults to its own module.
